+++ Common/Json/JsonWriter.cs
@@ -67,15 +67,15 @@ namespace Ks
 
                 if ((this.MultiLine & !this.HasKeyBefore) | NewLineRequired)
                 {
-                    if ((int)this.State != (int)WriterState.Begin)
+                    if (this.State != WriterState.Begin)
                         this.WriteNewLine();
                 }
             }
 
             public void WriteValue(string Value, bool Quoted)
             {
-                Verify.False((int)this.State == (int)WriterState.End, "Cannot write after write is finished.");
-                Verify.True(((int)this.State == (int)WriterState.Dictionary).Implies(this.HasKeyBefore), $"Cannot write a value in place of a key in a dictionary. Use {nameof(this.WriteKey)} instead.");
+                Verify.False(this.State == WriterState.End, "Cannot write after write is finished.");
+                Verify.True((this.State == WriterState.Dictionary).Implies(this.HasKeyBefore), $"Cannot write a value in place of a key in a dictionary. Use {nameof(this.WriteKey)} instead.");
 
                 this.WriteSeparator();
 
@@ -88,7 +88,7 @@ namespace Ks
                 else
                     this.Out.Write(Value);
 
-                if ((int)this.State == (int)WriterState.Begin)
+                if (this.State == WriterState.Begin)
                     this.State = WriterState.End;
                 this.HasKeyBefore = false;
                 this.HasValueBefore = true;
@@ -96,8 +96,8 @@ namespace Ks
 
             public void WriteKey(string Name)
             {
-                Verify.False((int)this.State == (int)WriterState.End, "Cannot write after write is finished.");
-                Verify.True((int)this.State == (int)WriterState.Dictionary, "Cannot write a key outside a dictionary.");
+                Verify.False(this.State == WriterState.End, "Cannot write after write is finished.");
+                Verify.True(this.State == WriterState.Dictionary, "Cannot write a key outside a dictionary.");
                 Verify.False(this.HasKeyBefore, "Cannot write a key immediately after another.");
 
                 this.WriteSeparator();
@@ -112,10 +112,10 @@ namespace Ks
 
             public Opening OpenList(bool MultiLine = false)
             {
-                Verify.False((int)this.State == (int)WriterState.End, "Cannot write after write is finished.");
-                Verify.True(((int)this.State == (int)WriterState.Dictionary).Implies(this.HasKeyBefore), $"Cannot write a value in place of a key in a dictionary. Use {nameof(this.WriteKey)} instead.");
+                Verify.False(this.State == WriterState.End, "Cannot write after write is finished.");
+                Verify.True((this.State == WriterState.Dictionary).Implies(this.HasKeyBefore), $"Cannot write a value in place of a key in a dictionary. Use {nameof(this.WriteKey)} instead.");
 
-                var R = new Opening(this, ']', ((int)this.State == (int)WriterState.Begin) ? WriterState.End : this.State, this.MultiLine);
+                var R = new Opening(this, ']', (this.State == WriterState.Begin) ? WriterState.End : this.State, this.MultiLine);
 
                 this.WriteSeparator(this.OpeningBraceOnNewLine & MultiLine);
                 this.Out.Write('[');
@@ -133,10 +133,10 @@ namespace Ks
 
             public Opening OpenDictionary(bool MultiLine = false)
             {
-                Verify.False((int)this.State == (int)WriterState.End, "Cannot write after write is finished.");
-                Verify.True(((int)this.State == (int)WriterState.Dictionary).Implies(this.HasKeyBefore), $"Cannot write a value in place of a key in a dictionary. Use {nameof(this.WriteKey)} instead.");
+                Verify.False(this.State == WriterState.End, "Cannot write after write is finished.");
+                Verify.True((this.State == WriterState.Dictionary).Implies(this.HasKeyBefore), $"Cannot write a value in place of a key in a dictionary. Use {nameof(this.WriteKey)} instead.");
 
-                var R = new Opening(this, '}', ((int)this.State == (int)WriterState.Begin) ? WriterState.End : this.State, this.MultiLine);
+                var R = new Opening(this, '}', (this.State == WriterState.Begin) ? WriterState.End : this.State, this.MultiLine);
 
                 this.WriteSeparator(this.OpeningBraceOnNewLine & MultiLine);
                 this.Out.Write('{');
@@ -155,7 +155,7 @@ namespace Ks
             private void CloseOpening(char ClosingChar, WriterState PreviousState, bool PreviousMultiline)
             {
                 Verify.False(this.HasKeyBefore, "Cannot close while a key is pending its value.");
-                Verify.False((int)this.State == (int)WriterState.End, "Cannot write after write is finished.");
+                Verify.False(this.State == WriterState.End, "Cannot write after write is finished.");
 
                 if (this.MultiLine)
                 {
